feat: initial working version
This is a first "working" version of the `jilu` application.

There are still many things to do, and there are still some hard-coded
debug variables in the source code, but it's good to get it out there,
and iterate on it from here on out.

The three next big steps are:

1. Improve testing setup.
2. Expose configurables.
3. Expand documentation.

But it's out there, yay! 🎉 💃🏽