docs: add improved usage instructions
Specifically, mention the fact that Jiju has to read the configuration
from the same file it will be writing its output to. This unfortunately
prevents the use of output redirection on Unix, because it will clear
the file's contents before Jiju is able to read the configuration
settings.

A workaround is documented, which works fine for now. Perhaps it makes
sense to support an optional `--out` flag once the CLI is fleshed out
more.