feat: show commit type of unreleased changes
The goal is to keep the list of unreleased changes more succinct, but
still provide enough information to see the most important upcoming
changes at a glance.

With this change, the commit type is added to the single-line
description of each change, giving better insight into how relevant each
change is.

When breaking changes are added to the change log, that too should
somehow be conveyed in the upcoming changes list, so people can
anticipate these breaking changes by reading the change log.