fix: remove extra whitespace at end of commit
The git2 library adds a newline at the end of commit messages, even if
the message is a single line. This makes sense when printing it to the
screen, but not for our parser.

Our parser either accepts a single line commit:

    feat: my commit message

Or a multi-line commit with a blank line between the subject and the
body:

    feat: my commit message

    with a commit body!

In the first case, git2 adds a newline after the subject, which falls
between the first and second case, and is thus considered invalid.

The solution is to always trim any extra whitespace at the end of the
commit message.