feat(config): try to get default remote origin url
With this commit, we try to get the default remote origin URL from Git's
`origin` remote.

If found, we use it to set the `github.repo` field in the configuration.

This ensures a more useful default usage of the tool, as it will
generate correct links when running `jilu` without any custom
configuration in a repository with a remote named `origin`.

Signed-off-by: Jean Mertz <git@jeanmertz.com>