fix(template): Github-independent anchor links for releases
Before this commit, we relied on Github's automatic anchor links to
allow clicking on releases in the "overview" section of the change log.

Even on Github however, those links were broken for titled releases, as
the anchor links would include the sanitized title of the release.

This commit addresses both issues; the anchor links are now Github-
independent, and the anchor links now work for titled and untitled
releases.

Signed-off-by: Jean Mertz <git@jeanmertz.com>