feat: Show unreleased changes in release notes editor
When creating a release with the `RELEASE` environment variable, if
`RELEASE_EDIT` is set, the unreleased commit messages are now displayed
in the release notes editor as comments. This provides immediate context
about what changes are being included in the release, making it more
convenient to write release notes without having to reference the git
log separately.

Signed-off-by: Jean Mertz <git@jeanmertz.com>