feat: Optionally write change log to file
Previously, `jilu` would always write the change log to stdout, allowing
you to pipe it to a file if needed. This caused a few issues:

- You needed to use `sponge` to keep the inline configuration in the
  relevant change log file.
- Opening `$EDITOR` does not work reliably when piping stdout.

With this change, those issues are resolved.

The previous workflow was:

```sh
jilu | sponge CHANGELOG.md
```

The new is:

```sh
jilu --write CHANGELOG.md
```

- You provide the change log file as the last argument to `jilu`.
- You (optionally) provide the `--write` flag to let `jilu` write the
  changes to the file instead of printing them to stdout.

This removes the need for `sponge`, and allows the editor to be used
when `RELEASE_EDIT` is set.

You can still run `jilu` without any arguments, which reads
`CHANGELOG.md` for any custom configurations, if it exists, and prints
the result to stdout.

Signed-off-by: Jean Mertz <git@jeanmertz.com>