feat: Support contributor footers in changelog generation
This change enhances contributor tracking by expanding beyond just
commit authors to include contributors listed in commit footers. The
system now recognizes common footer tokens like `Co-authored-by`,
`Signed-off-by`, `Reviewed-by`, and others to create a comprehensive
list of contributors for each change.

The list of contributor footers can be configured using the
`contributor_footers` configuration option.

The list defaults to:

- co-authored-by
- signed-off-by
- reported-by
- tested-by
- reviewed-by
- suggested-by
- acked-by

The author and committer of a change are always included in the list of
contributors, contributors are sorted alphabetically, and duplicate
entries are removed.

Reviewed-by: Homer Simpson <homer@simpsons.com>
Acked-by: Marge Simpson <marge@simpsons.com>
Tested-by: Bart Simpson <bart@simpsons.com>
Suggested-by: Lisa Simpson <lisa@simpsons.com>
Signed-off-by: Jean Mertz <git@jeanmertz.com>